+++ src/compas/geometry/_predicates/__init__.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from .predicates_2 import *  # noqa: F401 F403
+from .predicates_3 import *  # noqa: F401 F403
+
+__all__ = [name for name in dir() if not name.startswith('_')]