+++ src/compas_blender/__init__.py
@@ -41,7 +41,7 @@ def clear():
             bpy.data.collections.remove(block)
 
 
-__version__ = '0.17.2'
+__version__ = '0.18.0'
 
 
 __all__ = [name for name in dir() if not name.startswith('_')]