+++ src/compas_rhino/artists/volmeshartist.py
@@ -184,7 +184,7 @@ class VolMeshArtist(BaseArtist):
         facets = []
         for face in faces:
             facets.append({
-                'points': [vertex_xyz[vertex] for vertex in self.volmesh.face_vertices(face)],
+                'points': [vertex_xyz[vertex] for vertex in self.volmesh.halfface_vertices(face)],
                 'name': "{}.face.{}".format(self.volmesh.name, face),
                 'color': face_color[face]})
         return compas_rhino.draw_faces(facets, layer=self.layer, clear=False, redraw=False)